Revert "jjgkk"
This reverts commit e717b5f48a9763e5bac09565c06f32b3b2b5ec94.

+++ src/main/java/com/to/t1/board/comment/CommentService.java
@@ -19,17 +19,17 @@ public class CommentService {
         return commentMapper.commentList(commentVO);
     }
     
-    public long commentInsert(CommentVO commentVO) throws Exception{
+    public int commentInsert(CommentVO commentVO) throws Exception{
         
         return commentMapper.commentInsert(commentVO);
     }
     
-    public long commentUpdate(CommentVO commentVO) throws Exception{
+    public int commentUpdate(CommentVO commentVO) throws Exception{
         
         return commentMapper.commentUpdate(commentVO);
     }
     
-    public long commentDelete(long coNum) throws Exception{
+    public int commentDelete(long coNum) throws Exception{
         
         return commentMapper.commentDelete(coNum);
     }
